Refactor the library out into multiple files
One main 1300-line module is difficult to navigate within. This commit
splits that out into individual modules roughly broken down by their
main type. Specifically, we have:
- a module for Group,
- a module for Counter,
- a module for Sampler,
- a module for Builder,
- and a module for any flags (currently just SampleFlags)

This should make it easier to find whatever code is relevant in the
future.
